Replace the old implementation with the new TDD one!
